Extract Ordering tests
